server/tests-py/requirements.txt
aniso8601==6.0.0
apipkg==1.5
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
dirsync==2.2.4
docker==4.0.2
execnet==1.7.0
graphene==2.1.7
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
importlib-metadata==0.19
Jinja2==2.10.1
jsondiff==1.2.0
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
promise==2.2.1
psutil==5.6.3
psycopg2-binary==2.8.3
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.0.1
pytest-forked==1.0.2
pytest-xdist==1.29.0
PyYAML==5.1.2
requests==2.22.0
Rx==1.6.1
sh==1.12.14
six==1.12.0
SQLAlchemy==1.3.6
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
zipp==0.5.2
